pyproject.toml
[project]
name = "amzn-fkat"
authors = [
  { name="FKAT Contributors" },
]
description = "Foundational Kit for AI Training"
license = { text="Apache-2.0" }
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = [
  "hydra-core",
  "torch>=2.0.1,!=2.3.0,!=2.6.0",  # excluding 2.3.0 because of a breaking change: https://github.com/pytorch/pytorch/pull/122616
  "torchvision",  # has to be installed together with pytorch to be compatible
  "torchaudio",  # has to be installed together with pytorch to be compatible
  "torchmetrics>=0.11.4",
  "lightning!=2022.*",
  "lightning-utilities",
  "transformers",
  "boto3>=1.35.89",
  "fsspec[s3]",
  "importlib-metadata",
  "datasets>=3.0.0",
  "evaluate",
  "awswrangler>=3.5.1",
  "pandas",
  "mlflow-skinny",
  "pyarrow>=15.0.0,<21.0.0",  # Use stable versions with pre-built wheels
  "nvidia-ml-py",
]

[project.urls]
homepage = "https://github.com/amzn/fkat"

[project.optional-dependencies]
test = [
  "nvidia-ml-py",
  "ruff",
  "pre-commit",
  "pre-commit-hooks",
  "captum>=0.4.0",
  "contourpy<1.3.1",
  "deepspeed",
  "nvtx",
  "viztracer",
  "ipython==8.18.1",
  "moto==5.0.18",
  "trl==0.7.1",
  "peft",
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "pytest-timeout",
  "ty",
  "tox-venv",
  "tox==3.26.0",
  "tox-wheel==1.0.0",
  "twine",
  "types-boto3[batch,dynamodb,ec2,s3]",
]
docs = [
  "sphinx>5.0,<6.0",
  "myst-parser>=0.18.1,<3.0.0",
  "pandoc>=1.0,<=2.3",
  "docutils>=0.16,<0.21",
  "sphinxcontrib-mockautodoc",
  "sphinx-autobuild",
  "sphinx-autodoc-typehints==1.23.4",
  "sphinx-paramlinks>=0.5.1,<=0.6.0",
  "sphinx-togglebutton>=0.2,<=0.3.2",
  "sphinx-copybutton>=0.3,<=0.5.2",
  "sphinx-multiproject",
  "sphinx-toolbox==3.5.0",
  "sphinx-prompt==1.5.0",
  "sphinx-rtd-dark-mode==1.3.0",
  "sphinxcontrib-video==0.2.0",
  "sphinx-book-theme",
]

# Hatch
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
targets.wheel.include = ["py.typed"]

[tool.hatch.version]
path = "fkat/__version__.py"

[tool.hatch.build]
directory = "dist"
exclude = [
  "**/test/**"
]

[tool.hatch.build.targets.wheel]
packages = ["fkat"]

[tool.hatch.env]
requires = [
  "hatch-pip-compile",
  "virtualenv<20.35.0",  # https://github.com/pypa/hatch/issues/2079
  "pip<25.2",  # https://github.com/jazzband/pip-tools/issues/2252
]

[tool.uv.workspace]

[tool.hatch.env.collectors.custom]
path = ".hatch/hatch_plugin.py"

[tool.coverage.run]
branch = true
omit = [
    "*/_remote_module_non_scriptable.py",  # omit pytorch-generated files in /tmp
]
[tool.coverage.xml]
output = "coverage.xml"
[tool.coverage.html]
directory = "htmlcov/"

[tool.coverage.report]
fail_under = 86
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
  "__pycache__",
  ".env",
  "dist",
  "build",
  ".venv",
  ".ty_cache",
  ".git",
  ".tox",
]

[tool.ruff.lint]
# B: flake8-bugbear
# C: flake8-comprehensions
# E: pycodestyle errors
# F: pyflakes
# W: pycodestyle warnings
# I: isort
# N: pep8-naming
# D: pydocstyle
# UP: pyupgrade
# B9: flake8-broken-line
# ANN: flake8-annotations
# BLE: flake8-blind-except
# A: flake8-builtins
# COM: flake8-commas
# SIM: flake8-simplify
# ARG: flake8-unused-arguments
# ERA: eradicate
# PD: pandas-vet
# PL: pylint
# RUF: ruff-specific rules
select = ["B", "C", "E", "F", "W", "B9", "UP", "ANN"]
ignore = ["ANN401"]  # Allow Any type

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["fkat"]
force-single-line = false
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
"**/test/**" = ["D", "ANN"]
"**/__init__.py" = ["F401"]

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.ty]

